Add "special" objective lists.
Instead of specifying an array of objectives in a layout, one can now
specify a special tag.  Currently that is only "checks" which will
cause the objectives to the populated with the active pop-up's
checks.  This is in support of modules being able to supply
their own check pop-up layouts.